+++ packages/zero-pg/src/connection.ts
@@ -1,0 +1,7 @@
+export interface Connection {
+  transact<R>(callback: (tx: ConnectionTransaction) => Promise<R>): Promise<R>;
+}
+
+export interface ConnectionTransaction {
+  query(sql: string, params: unknown[]): Promise<unknown[]>;
+}
