+++ packages/zero-pg/src/mod.ts
@@ -1,9 +1,3 @@
-export {
-  makeServerTransaction,
-  makeSchemaCRUD,
-  type CustomMutatorDefs,
-  type CustomMutatorImpl,
-} from './custom.ts';
 export {makeSchemaQuery} from './query.ts';
 export type {Row} from '../../zql/src/mutate/custom.ts';
 export {ZQLPGDatabase} from './zql-pg-database.ts';
